@@ -3,10 +3,9 @@
   <PropertyGroup>
     <Copyright>2015-2021 The Neo Project</Copyright>
     <AssemblyTitle>Neo</AssemblyTitle>
-    <VersionPrefix>3.0.0</VersionPrefix>
-    <VersionSuffix>rc1</VersionSuffix>
+    <VersionPrefix>3.1.0</VersionPrefix>
     <Authors>The Neo Project</Authors>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <AssemblyName>Neo</AssemblyName>
     <PackageId>Neo</PackageId>
@@ -26,11 +25,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Akka" Version="1.4.17" />
+    <PackageReference Include="Akka" Version="1.4.28" />
     <PackageReference Include="BouncyCastle.NetCore" Version="1.8.8" />
-    <PackageReference Include="K4os.Compression.LZ4" Version="1.2.6" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="5.0.4" />
-    <PackageReference Include="Neo.VM" Version="3.0.0-CI00272" />
+    <PackageReference Include="K4os.Compression.LZ4" Version="1.2.15" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="6.0.0" />
+    <PackageReference Include="Neo.VM" Version="3.1.0" />
   </ItemGroup>
 
 </Project>
