@@ -1,3 +1,13 @@
+// Copyright (C) 2015-2021 The Neo Project.
+// 
+// The neo is free software distributed under the MIT software license, 
+// see the accompanying file LICENSE in the main directory of the
+// project or http://www.opensource.org/licenses/mit-license.php 
+// for more details.
+// 
+// Redistribution and use in source and binary forms with or without
+// modifications are permitted.
+
 using System;
 using System.Numerics;
 
@@ -6,7 +16,7 @@ namespace Neo
     /// <summary>
     /// Represents a fixed-point number of arbitrary precision.
     /// </summary>
-    public struct BigDecimal
+    public struct BigDecimal : IComparable<BigDecimal>, IEquatable<BigDecimal>
     {
         private readonly BigInteger value;
         private readonly byte decimals;
@@ -43,9 +53,15 @@ namespace Neo
         /// <param name="value">The value of the number.</param>
         public unsafe BigDecimal(decimal value)
         {
-            ReadOnlySpan<byte> buffer = new(&value, sizeof(decimal));
-            this.decimals = buffer[14];
-            this.value = new BigInteger(decimal.Multiply((decimal)Math.Pow(10, decimals), value));
+            Span<int> span = stackalloc int[4];
+            decimal.GetBits(value, span);
+            fixed (int* p = span)
+            {
+                ReadOnlySpan<byte> buffer = new(p, 16);
+                this.value = new BigInteger(buffer[..12], isUnsigned: true);
+                if (buffer[15] != 0) this.value = -this.value;
+                this.decimals = buffer[14];
+            }
         }
 
         /// <summary>
@@ -55,9 +71,19 @@ namespace Neo
         /// <param name="decimals">The number of decimal places for this number.</param>
         public unsafe BigDecimal(decimal value, byte decimals)
         {
-            ReadOnlySpan<byte> buffer = new(&value, sizeof(decimal));
-            if (buffer[14] > decimals) throw new ArgumentException(null, nameof(value));
-            this.value = new BigInteger(decimal.Multiply((decimal)Math.Pow(10, decimals), value));
+            Span<int> span = stackalloc int[4];
+            decimal.GetBits(value, span);
+            fixed (int* p = span)
+            {
+                ReadOnlySpan<byte> buffer = new(p, 16);
+                this.value = new BigInteger(buffer[..12], isUnsigned: true);
+                if (buffer[14] > decimals)
+                    throw new ArgumentException(null, nameof(value));
+                else if (buffer[14] < decimals)
+                    this.value *= BigInteger.Pow(10, decimals - buffer[14]);
+                if (buffer[15] != 0)
+                    this.value = -this.value;
+            }
             this.decimals = decimals;
         }
 
@@ -153,6 +179,64 @@ namespace Neo
             }
             result = new BigDecimal(value, decimals);
             return true;
+        }
+
+        public int CompareTo(BigDecimal other)
+        {
+            BigInteger left = value, right = other.value;
+            if (decimals < other.decimals)
+                left *= BigInteger.Pow(10, other.decimals - decimals);
+            else if (decimals > other.decimals)
+                right *= BigInteger.Pow(10, decimals - other.decimals);
+            return left.CompareTo(right);
+        }
+
+        public override bool Equals(object obj)
+        {
+            if (obj is not BigDecimal @decimal) return false;
+            return Equals(@decimal);
+        }
+
+        public bool Equals(BigDecimal other)
+        {
+            return CompareTo(other) == 0;
+        }
+
+        public override int GetHashCode()
+        {
+            BigInteger divisor = BigInteger.Pow(10, decimals);
+            BigInteger result = BigInteger.DivRem(value, divisor, out BigInteger remainder);
+            return HashCode.Combine(result, remainder);
+        }
+
+        public static bool operator ==(BigDecimal left, BigDecimal right)
+        {
+            return left.CompareTo(right) == 0;
+        }
+
+        public static bool operator !=(BigDecimal left, BigDecimal right)
+        {
+            return left.CompareTo(right) != 0;
+        }
+
+        public static bool operator <(BigDecimal left, BigDecimal right)
+        {
+            return left.CompareTo(right) < 0;
+        }
+
+        public static bool operator <=(BigDecimal left, BigDecimal right)
+        {
+            return left.CompareTo(right) <= 0;
+        }
+
+        public static bool operator >(BigDecimal left, BigDecimal right)
+        {
+            return left.CompareTo(right) > 0;
+        }
+
+        public static bool operator >=(BigDecimal left, BigDecimal right)
+        {
+            return left.CompareTo(right) >= 0;
         }
     }
 }
