@@ -41,7 +41,7 @@ namespace Neo.UnitTests.Wallets.NEP6
                 rng.GetBytes(privateKey);
             }
             keyPair = new KeyPair(privateKey);
-            testScriptHash = Neo.SmartContract.Contract.CreateSignatureContract(keyPair.PublicKey).ScriptHash;
+            testScriptHash = Contract.CreateSignatureContract(keyPair.PublicKey).ScriptHash;
             nep2key = keyPair.Export("123", ProtocolSettings.Default.AddressVersion, 2, 1, 1);
         }
 
@@ -96,9 +96,9 @@ namespace Neo.UnitTests.Wallets.NEP6
         [TestMethod]
         public void TestChangePassword()
         {
-            JObject wallet = new JObject();
+            JObject wallet = new();
             wallet["name"] = "name";
-            wallet["version"] = new System.Version("1.0").ToString();
+            wallet["version"] = new Version("1.0").ToString();
             wallet["scrypt"] = new ScryptParameters(2, 1, 1).ToJson();
             wallet["accounts"] = new JArray();
             wallet["extra"] = new JObject();
@@ -116,7 +116,7 @@ namespace Neo.UnitTests.Wallets.NEP6
         [TestMethod]
         public void TestConstructorWithPathAndName()
         {
-            NEP6Wallet wallet = new NEP6Wallet(wPath, ProtocolSettings.Default);
+            NEP6Wallet wallet = new(wPath, ProtocolSettings.Default);
             Assert.AreEqual("name", wallet.Name);
             Assert.AreEqual(new ScryptParameters(2, 1, 1).ToJson().ToString(), wallet.Scrypt.ToJson().ToString());
             Assert.AreEqual(new Version("1.0").ToString(), wallet.Version.ToString());
@@ -129,14 +129,14 @@ namespace Neo.UnitTests.Wallets.NEP6
         [TestMethod]
         public void TestConstructorWithJObject()
         {
-            JObject wallet = new JObject();
+            JObject wallet = new();
             wallet["name"] = "test";
             wallet["version"] = Version.Parse("1.0").ToString();
             wallet["scrypt"] = ScryptParameters.Default.ToJson();
             wallet["accounts"] = new JArray();
             wallet["extra"] = new JObject();
             wallet.ToString().Should().Be("{\"name\":\"test\",\"version\":\"1.0\",\"scrypt\":{\"n\":16384,\"r\":8,\"p\":8},\"accounts\":[],\"extra\":{}}");
-            NEP6Wallet w = new NEP6Wallet(null, ProtocolSettings.Default, wallet);
+            NEP6Wallet w = new(null, ProtocolSettings.Default, wallet);
             Assert.AreEqual("test", w.Name);
             Assert.AreEqual(Version.Parse("1.0").ToString(), w.Version.ToString());
         }
@@ -150,7 +150,7 @@ namespace Neo.UnitTests.Wallets.NEP6
         [TestMethod]
         public void TestGetVersion()
         {
-            Assert.AreEqual(new System.Version("1.0").ToString(), uut.Version.ToString());
+            Assert.AreEqual(new Version("1.0").ToString(), uut.Version.ToString());
         }
 
         [TestMethod]
@@ -264,14 +264,14 @@ namespace Neo.UnitTests.Wallets.NEP6
         [TestMethod]
         public void TestGetAccounts()
         {
-            Dictionary<UInt160, KeyPair> keys = new Dictionary<UInt160, KeyPair>();
+            Dictionary<UInt160, KeyPair> keys = new();
             uut.Unlock("123");
             byte[] privateKey = new byte[32];
             using (RandomNumberGenerator rng = RandomNumberGenerator.Create())
             {
                 rng.GetBytes(privateKey);
             }
-            KeyPair key = new KeyPair(privateKey);
+            KeyPair key = new(privateKey);
             keys.Add(Contract.CreateSignatureRedeemScript(key.PublicKey).ToScriptHash(), key);
             keys.Add(Contract.CreateSignatureRedeemScript(keyPair.PublicKey).ToScriptHash(), keyPair);
             uut.CreateAccount(key.PrivateKey);
@@ -288,7 +288,7 @@ namespace Neo.UnitTests.Wallets.NEP6
         public X509Certificate2 NewCertificate()
         {
             ECDsa key = ECDsa.Create(ECCurve.NamedCurves.nistP256);
-            CertificateRequest request = new CertificateRequest(
+            CertificateRequest request = new(
                 "CN=Self-Signed ECDSA",
                 key,
                 HashAlgorithmName.SHA256);
@@ -335,7 +335,7 @@ namespace Neo.UnitTests.Wallets.NEP6
             uut.DeleteAccount(testScriptHash);
             result = uut.Contains(testScriptHash);
             Assert.AreEqual(false, result);
-            JObject wallet = new JObject();
+            JObject wallet = new();
             wallet["name"] = "name";
             wallet["version"] = new Version("1.0").ToString();
             wallet["scrypt"] = new ScryptParameters(2, 1, 1).ToJson();
@@ -372,16 +372,16 @@ namespace Neo.UnitTests.Wallets.NEP6
             NEP6Wallet nw = NEP6Wallet.Migrate(npath, path, "123", ProtocolSettings.Default);
             bool result = nw.Contains(testScriptHash);
             Assert.AreEqual(true, result);
-            if (File.Exists(path)) File.Delete(path);
-            if (File.Exists(npath)) File.Delete(npath);
+            uw.Delete();
+            nw.Delete();
         }
 
         [TestMethod]
         public void TestSave()
         {
-            JObject wallet = new JObject();
+            JObject wallet = new();
             wallet["name"] = "name";
-            wallet["version"] = new System.Version("1.0").ToString();
+            wallet["version"] = new Version("1.0").ToString();
             wallet["scrypt"] = new ScryptParameters(2, 1, 1).ToJson();
             wallet["accounts"] = new JArray();
             wallet["extra"] = new JObject();
@@ -421,7 +421,7 @@ namespace Neo.UnitTests.Wallets.NEP6
             Assert.AreEqual(true, result);
             uut.DeleteAccount(testScriptHash);
             Assert.AreEqual(false, uut.Contains(testScriptHash));
-            JObject wallet = new JObject();
+            JObject wallet = new();
             wallet["name"] = "name";
             wallet["version"] = new Version("1.0").ToString();
             wallet["scrypt"] = new ScryptParameters(2, 1, 1).ToJson();
@@ -441,6 +441,25 @@ namespace Neo.UnitTests.Wallets.NEP6
             uut.Version.Should().Be(new Version("1.0"));
             uut.Scrypt.Should().NotBeNull();
             uut.Scrypt.N.Should().Be(new ScryptParameters(2, 1, 1).N);
+        }
+
+        [TestMethod]
+        public void TestIsDefault()
+        {
+            JObject wallet = new();
+            wallet["name"] = "name";
+            wallet["version"] = new Version("1.0").ToString();
+            wallet["scrypt"] = new ScryptParameters(2, 1, 1).ToJson();
+            wallet["accounts"] = new JArray();
+            wallet["extra"] = new JObject();
+            var w = new NEP6Wallet(null, ProtocolSettings.Default, wallet);
+            using var l = w.Unlock("");
+            var ac = w.CreateAccount();
+            Assert.AreEqual(ac.Address, w.GetDefaultAccount().Address);
+            var ac2 = w.CreateAccount();
+            Assert.AreEqual(ac.Address, w.GetDefaultAccount().Address);
+            ac2.IsDefault = true;
+            Assert.AreEqual(ac2.Address, w.GetDefaultAccount().Address);
         }
     }
 }
