@@ -1,3 +1,13 @@
+// Copyright (C) 2015-2021 The Neo Project.
+// 
+// The neo is free software distributed under the MIT software license, 
+// see the accompanying file LICENSE in the main directory of the
+// project or http://www.opensource.org/licenses/mit-license.php 
+// for more details.
+// 
+// Redistribution and use in source and binary forms with or without
+// modifications are permitted.
+
 using Neo.SmartContract;
 using System.Collections.Generic;
 
@@ -14,7 +24,7 @@ namespace Neo.Persistence
 
         protected override void AddInternal(StorageKey key, StorageItem value)
         {
-            innerCache.Add(key, value);
+            innerCache.Add(key, value.Clone());
         }
 
         protected override void DeleteInternal(StorageKey key)
