@@ -1,3 +1,13 @@
+// Copyright (C) 2015-2021 The Neo Project.
+// 
+// The neo is free software distributed under the MIT software license, 
+// see the accompanying file LICENSE in the main directory of the
+// project or http://www.opensource.org/licenses/mit-license.php 
+// for more details.
+// 
+// Redistribution and use in source and binary forms with or without
+// modifications are permitted.
+
 using Neo.Cryptography;
 using Neo.Cryptography.ECC;
 using Neo.Network.P2P.Payloads;
@@ -21,7 +31,7 @@ namespace Neo.SmartContract
         /// <summary>
         /// The maximum GAS that can be consumed when <see cref="VerifyWitnesses"/> is called.
         /// </summary>
-        public const long MaxVerificationGas = 0_50000000;
+        public const long MaxVerificationGas = 1_50000000;
 
         /// <summary>
         /// Calculates the verification fee for a signature address.
