@@ -1,3 +1,13 @@
+// Copyright (C) 2015-2021 The Neo Project.
+// 
+// The neo is free software distributed under the MIT software license, 
+// see the accompanying file LICENSE in the main directory of the
+// project or http://www.opensource.org/licenses/mit-license.php 
+// for more details.
+// 
+// Redistribution and use in source and binary forms with or without
+// modifications are permitted.
+
 using Neo.Cryptography.ECC;
 using Neo.Network.P2P.Payloads;
 
@@ -18,7 +28,7 @@ namespace Neo.SmartContract.Native
         internal override ContractTask Initialize(ApplicationEngine engine)
         {
             UInt160 account = Contract.GetBFTAddress(engine.ProtocolSettings.StandbyValidators);
-            return Mint(engine, account, 30_000_000 * Factor, false);
+            return Mint(engine, account, engine.ProtocolSettings.InitialGasDistribution, false);
         }
 
         internal override async ContractTask OnPersist(ApplicationEngine engine)
