@@ -1,3 +1,13 @@
+// Copyright (C) 2015-2021 The Neo Project.
+// 
+// The neo is free software distributed under the MIT software license, 
+// see the accompanying file LICENSE in the main directory of the
+// project or http://www.opensource.org/licenses/mit-license.php 
+// for more details.
+// 
+// Redistribution and use in source and binary forms with or without
+// modifications are permitted.
+
 using Neo.Cryptography;
 using Neo.Cryptography.ECC;
 using Neo.IO;
@@ -22,7 +32,7 @@ namespace Neo.SmartContract.Native
             {
                 new ContractEventDescriptor
                 {
-                    Name = "Desingation",
+                    Name = "Designation",
                     Parameters = new ContractParameterDefinition[]
                     {
                         new ContractParameterDefinition()
