@@ -1,3 +1,13 @@
+// Copyright (C) 2015-2021 The Neo Project.
+// 
+// The neo is free software distributed under the MIT software license, 
+// see the accompanying file LICENSE in the main directory of the
+// project or http://www.opensource.org/licenses/mit-license.php 
+// for more details.
+// 
+// Redistribution and use in source and binary forms with or without
+// modifications are permitted.
+
 using Neo.IO;
 using Neo.Network.P2P.Payloads;
 using Neo.VM;
@@ -20,6 +30,11 @@ namespace Neo.SmartContract.Native
         /// </summary>
         public Transaction Transaction;
 
+        /// <summary>
+        /// The execution state
+        /// </summary>
+        public VMState State;
+
         private StackItem _rawTransaction;
 
         void IInteroperable.FromStackItem(StackItem stackItem)
@@ -28,12 +43,13 @@ namespace Neo.SmartContract.Native
             BlockIndex = (uint)@struct[0].GetInteger();
             _rawTransaction = @struct[1];
             Transaction = _rawTransaction.GetSpan().AsSerializable<Transaction>();
+            State = (VMState)(byte)@struct[2].GetInteger();
         }
 
         StackItem IInteroperable.ToStackItem(ReferenceCounter referenceCounter)
         {
             _rawTransaction ??= Transaction.ToArray();
-            return new Struct(referenceCounter) { BlockIndex, _rawTransaction };
+            return new Struct(referenceCounter) { BlockIndex, _rawTransaction, (byte)State };
         }
     }
 }
