@@ -24,17 +24,17 @@ namespace Neo.UnitTests.Network.P2P.Payloads
         public void Size_Get()
         {
             UInt256 val256 = UInt256.Zero;
-            TestUtils.SetupHeaderWithValues(uut, val256, out _, out _, out _, out _, out _);
+            TestUtils.SetupHeaderWithValues(uut, val256, out _, out _, out _, out _, out _, out _);
             // blockbase 4 + 64 + 1 + 32 + 4 + 4 + 20 + 4
             // header 1
-            uut.Size.Should().Be(105);
+            uut.Size.Should().Be(113); // 105 + nonce
         }
 
         [TestMethod]
         public void GetHashCodeTest()
         {
             UInt256 val256 = UInt256.Zero;
-            TestUtils.SetupHeaderWithValues(uut, val256, out _, out _, out _, out _, out _);
+            TestUtils.SetupHeaderWithValues(uut, val256, out _, out _, out _, out _, out _, out _);
             uut.GetHashCode().Should().Be(uut.Hash.GetHashCode());
         }
 
@@ -43,7 +43,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
         {
             UInt256 val256 = UInt256.Zero;
             var snapshot = TestBlockchain.GetTestSnapshot().CreateSnapshot();
-            TestUtils.SetupHeaderWithValues(uut, val256, out _, out _, out _, out _, out _);
+            TestUtils.SetupHeaderWithValues(uut, val256, out _, out _, out _, out _, out _, out _);
             uut.Witness = new Witness() { InvocationScript = Array.Empty<byte>(), VerificationScript = Array.Empty<byte>() };
 
             UT_SmartContractHelper.BlocksAdd(snapshot, uut.Hash, new TrimmedBlock()
@@ -76,25 +76,26 @@ namespace Neo.UnitTests.Network.P2P.Payloads
         public void Deserialize()
         {
             UInt256 val256 = UInt256.Zero;
-            TestUtils.SetupHeaderWithValues(new Header(), val256, out UInt256 merkRoot, out UInt160 val160, out ulong timestampVal, out uint indexVal, out Witness scriptVal);
+            TestUtils.SetupHeaderWithValues(new Header(), val256, out UInt256 merkRoot, out UInt160 val160, out ulong timestampVal, out ulong nonceVal, out uint indexVal, out Witness scriptVal);
 
             uut.MerkleRoot = merkRoot; // need to set for deserialise to be valid
 
-            var hex = "0000000000000000000000000000000000000000000000000000000000000000000000007227ba7b747f1a98f68679d4a98b68927646ab195a6f56b542ca5a0e6a412662e913ff854c0000000000000000000000000000000000000000000000000000000001000111";
+            var hex = "0000000000000000000000000000000000000000000000000000000000000000000000007227ba7b747f1a98f68679d4a98b68927646ab195a6f56b542ca5a0e6a412662e913ff854c00000000000000000000000000000000000000000000000000000000000000000000000001000111";
 
             using MemoryStream ms = new(hex.HexToBytes(), false);
             using BinaryReader reader = new(ms);
             uut.Deserialize(reader);
 
-            AssertStandardHeaderTestVals(val256, merkRoot, val160, timestampVal, indexVal, scriptVal);
+            AssertStandardHeaderTestVals(val256, merkRoot, val160, timestampVal, nonceVal, indexVal, scriptVal);
         }
 
-        private void AssertStandardHeaderTestVals(UInt256 val256, UInt256 merkRoot, UInt160 val160, ulong timestampVal, uint indexVal, Witness scriptVal)
+        private void AssertStandardHeaderTestVals(UInt256 val256, UInt256 merkRoot, UInt160 val160, ulong timestampVal, ulong nonceVal, uint indexVal, Witness scriptVal)
         {
             uut.PrevHash.Should().Be(val256);
             uut.MerkleRoot.Should().Be(merkRoot);
             uut.Timestamp.Should().Be(timestampVal);
             uut.Index.Should().Be(indexVal);
+            uut.Nonce.Should().Be(nonceVal);
             uut.NextConsensus.Should().Be(val160);
             uut.Witness.InvocationScript.Length.Should().Be(0);
             uut.Witness.Size.Should().Be(scriptVal.Size);
@@ -119,8 +120,8 @@ namespace Neo.UnitTests.Network.P2P.Payloads
         {
             Header newHeader = new();
             UInt256 prevHash = new(TestUtils.GetByteArray(32, 0x42));
-            TestUtils.SetupHeaderWithValues(newHeader, prevHash, out _, out _, out _, out _, out _);
-            TestUtils.SetupHeaderWithValues(uut, prevHash, out _, out _, out _, out _, out _);
+            TestUtils.SetupHeaderWithValues(newHeader, prevHash, out _, out _, out _, out _, out _, out _);
+            TestUtils.SetupHeaderWithValues(uut, prevHash, out _, out _, out _, out _, out _, out _);
 
             uut.Equals(newHeader).Should().BeTrue();
         }
@@ -135,9 +136,9 @@ namespace Neo.UnitTests.Network.P2P.Payloads
         public void Serialize()
         {
             UInt256 val256 = UInt256.Zero;
-            TestUtils.SetupHeaderWithValues(uut, val256, out _, out _, out _, out _, out _);
+            TestUtils.SetupHeaderWithValues(uut, val256, out _, out _, out _, out _, out _, out _);
 
-            var hex = "0000000000000000000000000000000000000000000000000000000000000000000000007227ba7b747f1a98f68679d4a98b68927646ab195a6f56b542ca5a0e6a412662e913ff854c0000000000000000000000000000000000000000000000000000000001000111";
+            var hex = "0000000000000000000000000000000000000000000000000000000000000000000000007227ba7b747f1a98f68679d4a98b68927646ab195a6f56b542ca5a0e6a412662e913ff854c00000000000000000000000000000000000000000000000000000000000000000000000001000111";
             uut.ToArray().ToHexString().Should().Be(hex);
         }
     }
