@@ -10,6 +10,7 @@ using Neo.SmartContract.Native;
 using Neo.VM;
 using Neo.Wallets;
 using System;
+using System.IO;
 using System.Linq;
 using System.Numerics;
 
@@ -707,7 +708,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
             var signers = new Signer[]{ new Signer
                 {
                     Account = acc.ScriptHash,
-                    Scopes = (WitnessScope) 0xFF,
+                    Scopes = WitnessScope.CustomContracts,
                     AllowedContracts = new[] { NativeContract.NEO.Hash, NativeContract.GAS.Hash }
                 } };
 
@@ -841,14 +842,16 @@ namespace Neo.UnitTests.Network.P2P.Payloads
             tx2.NetworkFee.Should().Be(0x0000000000000001);
             tx2.ValidUntilBlock.Should().Be(0x01020304);
             tx2.Attributes.Should().BeEquivalentTo(Array.Empty<TransactionAttribute>());
-            tx2.Signers.Should().BeEquivalentTo(new Signer[] {
-                new Signer()
+            tx2.Signers.Should().BeEquivalentTo(new[]
+            {
+                new Signer
                 {
                     Account = UInt160.Zero,
                     AllowedContracts = Array.Empty<UInt160>(),
-                    AllowedGroups = Array.Empty<ECPoint>() }
+                    AllowedGroups = Array.Empty<ECPoint>(),
+                    Rules = Array.Empty<WitnessRule>()
                 }
-            );
+            });
             tx2.Script.Should().BeEquivalentTo(new byte[] { (byte)OpCode.PUSH1 });
             tx2.Witnesses.Should().BeEquivalentTo(new Witness[] { new Witness() { InvocationScript = Array.Empty<byte>(), VerificationScript = Array.Empty<byte>() } });
         }
@@ -1140,7 +1143,7 @@ namespace Neo.UnitTests.Network.P2P.Payloads
                     }
                 }
             };
-            tx.VerifyStateIndependent(ProtocolSettings.Default).Should().Be(VerifyResult.Invalid);
+            tx.VerifyStateIndependent(ProtocolSettings.Default).Should().Be(VerifyResult.OverSize);
             tx.Script = Array.Empty<byte>();
             tx.VerifyStateIndependent(ProtocolSettings.Default).Should().Be(VerifyResult.Succeed);
 
@@ -1193,6 +1196,15 @@ namespace Neo.UnitTests.Network.P2P.Payloads
 
             tx.Witnesses = data.GetWitnesses();
             tx.VerifyStateIndependent(ProtocolSettings.Default).Should().Be(VerifyResult.Succeed);
+
+            // Different hash
+
+            tx.Witnesses[0] = new Witness()
+            {
+                VerificationScript = walletB.GetAccounts().First().Contract.Script,
+                InvocationScript = tx.Witnesses[0].InvocationScript.ToArray()
+            };
+            tx.VerifyStateIndependent(ProtocolSettings.Default).Should().Be(VerifyResult.Invalid);
         }
 
         [TestMethod]
@@ -1274,6 +1286,20 @@ namespace Neo.UnitTests.Network.P2P.Payloads
 
             tx.Witnesses = data.GetWitnesses();
             tx.VerifyStateDependent(ProtocolSettings.Default, snapshot, new TransactionVerificationContext()).Should().Be(VerifyResult.Succeed);
+        }
+
+        [TestMethod]
+        public void Test_VerifyStateInDependent_Multi()
+        {
+            var txData = Convert.FromBase64String(
+                "AHXd31W0NlsAAAAAAJRGawAAAAAA3g8CAAGSs5x3qmDym1fBc87ZF/F/0yGm6wEAXwsDAOQLVAIAAAAMFLqZBJj+L0XZPXNHHM9MBfCza5HnDBSSs5x3qmDym1fBc87ZF/F/0yGm6xTAHwwIdHJhbnNmZXIMFM924ovQBixKR47jVWEBExnzz6TSQWJ9W1I5Af1KAQxAnZvOQOCdkM+j22dS5SdEncZVYVVi1F26MhheNzNImTD4Ekw5kFR6Fojs7gD57Bdeuo8tLS1UXpzflmKcQ3pniAxAYvGgxtokrk6PVdduxCBwVbdfie+ZxiaDsjK0FYregl24cDr2v5cTLHrURVfJJ1is+4G6Jaer7nB1JrDrw+Qt6QxATA5GdR4rKFPPPQQ24+42OP2tz0HylG1LlANiOtIdag3ZPkUfZiBfEGoOteRD1O0UnMdJP4Su7PFhDuCdHu4MlwxAuGFEk2m/rdruleBGYz8DIzExJtwb/TsFxZdHxo4VV8ktv2Nh71Fwhg2bhW2tq8hV6RK2GFXNAU72KAgf/Qv6BQxA0j3srkwY333KvGNtw7ZvSG8X36Tqu000CEtDx4SMOt8qhVYGMr9PClsUVcYFHdrJaodilx8ewXDHNIq+OnS7SfwVDCEDAJt1QOEPJWLl/Y+snq7CUWaliybkEjSP9ahpJ7+sIqIMIQMCBenO+upaHfxYCvIMjVqiRouwFI8aXkYF/GIsgOYEugwhAhS68M7qOmbxfn4eg56iX9i+1s2C5rtuaCUBiQZfRP8BDCECPpsy6om5TQZuZJsST9UOOW7pE2no4qauGxHBcNAiJW0MIQNAjc1BY5b2R4OsWH6h4Vk8V9n+qIDIpqGSDpKiWUd4BgwhAqeDS+mzLimB0VfLW706y0LP0R6lw7ECJNekTpjFkQ8bDCECuixw9ZlvNXpDGYcFhZ+uLP6hPhFyligAdys9WIqdSr0XQZ7Q3Do=");
+
+            var tx = new Transaction();
+            ((ISerializable)tx).Deserialize(new BinaryReader(new MemoryStream(txData)));
+
+            var settings = new ProtocolSettings() { Network = 844378958 };
+            var result = tx.VerifyStateIndependent(settings);
+            Assert.AreEqual(VerifyResult.Succeed, result);
         }
     }
 }
